strategy-engine/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
pandas>=2.2
numpy>=1.26
pandas-ta>=0.3.14b
vectorbt>=0.26
scikit-learn>=1.4
lightgbm>=4.3
apscheduler>=3.10
duckdb>=0.10
alpaca-py>=0.21
yfinance>=0.2
httpx>=0.27
pydantic>=2.6
transformers>=4.38
torch>=2.2
pytest>=8.0
pytest-asyncio>=0.23
